made the gps error an int
The float handling code was throwing exceptions and thereby killing the
connection and this also reduces the produced traffic and should be
sufficiently accurate.

+++ android/src/de/teammeet/interfaces/IXMPPService.java
@@ -76,7 +76,7 @@ public interface IXMPPService {
 
 	void stopLocationTransmission();
 
-	void sendLocation(GeoPoint location, float accuracy) throws XMPPException;
+	void sendLocation(GeoPoint location, int accuracy) throws XMPPException;
 
 	void sendIndicator(GeoPoint location) throws XMPPException;
